backend-python/requirements.txt
Flask
mysql-connector-python
google-cloud-functions
google-cloud-run
requests
numpy
scikit-learn